@@ -1,7 +1,7 @@
 {
     "name": "artis-mosties/oauth-laravel",
     "type": "library",
-    "description": "OAuth Service Provider for Laravel 4",
+    "description": "OAuth Service Provider for Laravel 5",
     "keywords": ["OAuth", "Lusitanian", "laravel", "php"],
     "homepage": "https://github.com/artdarek/oauth-4-laravel",
     "license": "MIT",
@@ -19,11 +19,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3",
+        "php": ">=5.4",
         "artis-mosties/oauth": "0.3.*@dev"
     },
     "require-dev": {
-        "illuminate/support": "~4"
+        "illuminate/support": "~5"
     },
     "autoload": {
         "psr-0": {
