@@ -9,51 +9,52 @@ namespace Artdarek\OAuth;
 
 use Illuminate\Support\ServiceProvider;
 
-class OAuthServiceProvider extends ServiceProvider 
-{
+class OAuthServiceProvider extends ServiceProvider {
 
-    /**
-     * Indicates if loading of the provider is deferred.
-     *
-     * @var bool
-     */
-    protected $defer = false;
+	/**
+	 * Indicates if loading of the provider is deferred.
+	 *
+	 * @var bool
+	 */
+	protected $defer = false;
 
-    /**
-     * Bootstrap the application events.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-        $this->package('artdarek/oauth-4-laravel');
-    }
+	/**
+	 * Bootstrap the application events.
+	 *
+	 * @return void
+	 */
+	public function boot()
+	{
+		$this->publishes([
+			__DIR__ . '/../../config/config.php' => config_path('oauth-laravel.php'),
+		], 'config');
+	}
 
-    /**
-     * Register the service provider.
-     *
-     * @return void
-     */
-    public function register()
-    {
-	    // Register 'oauth'
-		    $this->app['oauth'] = $this->app->share(function($app)
-		    {
-                // create oAuth instance
-                	$oauth = new OAuth();
-        		// return oAuth instance
-		        	return $oauth;
-		    });
-    }
+	/**
+	 * Register the service provider.
+	 *
+	 * @return void
+	 */
+	public function register()
+	{
+		// Register 'oauth'
+		$this->app['oauth'] = $this->app->share(function($app)
+		{
+			// create oAuth instance
+			$oauth = new OAuth();
+			// return oAuth instance
+			return $oauth;
+		});
+	}
 
-    /**
-     * Get the services provided by the provider.
-     *
-     * @return array
-     */
-    public function provides()
-    {
-        return array();
-    }
+	/**
+	 * Get the services provided by the provider.
+	 *
+	 * @return array
+	 */
+	public function provides()
+	{
+		return [];
+	}
 
 }
